Move menu and location models to appropriate location
