+++ lib/screens/landing/landing.dart
@@ -2,6 +2,7 @@ import 'package:WildcatMobileOrder/services/auth.dart';
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:WildcatMobileOrder/screens/menu/menu.dart';
+import 'package:WildcatMobileOrder/models/location.dart';
 
 class Landing extends StatelessWidget {
   final AuthService _auth = AuthService();
@@ -96,32 +97,3 @@ class Landing extends StatelessWidget {
   }
 }
 
-/// Holds a list of all locations, and their meta data.
-class Locations {
-  List<LocationItem> locations;
-
-  Locations.fromSnapshot(DocumentSnapshot snapshot)
-      : locations = snapshot['meta'].map<LocationItem>((loc) {
-          return LocationItem.fromMap(loc);
-        }).toList();
-}
-
-/// Represents a single location
-class LocationItem {
-  final String openTime;
-  final String closeTime;
-  final String name;
-
-  LocationItem.fromMap(Map<String, dynamic> map)
-      : openTime = map['opentime'],
-        closeTime = map['closetime'],
-        name = map['name'];
-
-  String getOpenHours() {
-    return 'Opens at $openTime and closes at $closeTime';
-  }
-
-  String getOrderString() {
-    return 'Order from $name';
-  }
-}
